Change to making multiple simultaneous cURL requests
Previously we were making a request, handling the response, making the next request,
handling the response to that, and so on.

Now we make multiple requests at the same time, and handle each response upon arrival.

This should have the benefit of speeding up load-times, particularly on pages that
require many requests to Repology's API.