Start migrating sections over to javascript
This allows for the use of asynchronous requests, meaning that we can draw the
main outline of the table, and populate the versions as we get them.

Instead of having to wait for all requests to complete before rendering the page.

This should also allow for better handling of "HTTP 429 Too Many Requests" error.