Add support for reading the PyPI API for python modules
The PyPI (Python Package Index) is not tracked by Repology,
but contains too many packages to track manually.

Fortunately, the service has an API which can be queried.
Unfortunately, it sometimes uses an alternate code for a given package.
Work needs to be done to correctly handle this latter thing.